Silence warnings

+++ cpp/include/API.h
@@ -119,7 +119,7 @@ public:
     
     ~ParserSession();
     
-    void init(BufferAndLength bufAndLen, WolframLibraryData libData, ParserSessionPolicy policy, SourceConvention srcConvention, size_t tabWidth);
+    void init(BufferAndLength bufAndLen, WolframLibraryData libData, ParserSessionPolicy policy, SourceConvention srcConvention, uint32_t tabWidth);
     
     void deinit();
